remove luigi and move targets to luigi_targets
