+++ luigi_targets.py
@@ -1,8 +1,30 @@
+from shell import cmd_output
 from azure.core.exceptions import ResourceNotFoundError
 from connector import Connector
 
 
-class AzureBlobTarget(luigi.Target):
+class AzureACRImageTarget():
+    """
+    Luigi target task for docker images in Azure ACR repositories.
+    """
+    def __init__(self, repository, tag):
+        self.path = f"{repository}:{tag}"
+        self.tag = tag
+        self.repository = repository
+
+    def acr_image_exists(self, repository, tag, account="fandango"):
+        """
+        checks whether a docker image exists in an azure ACR repository
+        """
+        cmd = f"az acr repository show-tags  --name {account} --resource-group fandango --repository {repository}"
+        output = cmd_output(cmd)
+        return f'"{tag}"' in output
+
+    def exists(self):
+        return self.acr_image_exists(self.repository, self.tag)
+
+
+class AzureBlobTarget():
     """
     Luigi Target class supporting smart_open like uris:
     azure://container/some/path/file
