+++ requirements.txt
@@ -5,4 +5,3 @@ azure-kusto-data==1.0.3
 azure-storage-blob==12.6.0
 smart_open==3.0.0
 smart_open[azure]==3.0.0
-luigi==2.8.11
